Big Changes - Including in EditProfileAdapter
CallbackActivity is for use in accessing MainActivity as a callback.
Please use it if not too inconvenient. Use Map instead of HashMap
(program to interface and not implementation). Also, reading for Profile
and Program and writing for Program are working. Let's not use old
values of hashmaps for the time being. It leads to incompatible and
undefined profiles being linked with valid programs. It would be best
just to read them in on startup. Then, after startup, we will used
sharedpreferences cheekily to keep things current from a unified source.